Функция getFromApiFirstTime работает
