+++ JS/renderComments.js
@@ -1,8 +1,8 @@
-import { comments } from "./comments.js";
+// import { comments } from "./comments.js";
 // import { likeButtonsListeners, editButtonsListeners, replyToCommentListeners } from "./main.js";
 
-const renderComments = (elem, getList) => {
-    const commentsHtml = comments
+const renderComments = (data, elem, getList) => {
+    const commentsHtml = data
         .map((comment, index) => getList(comment, index))
         .join('');
